% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disdRo_functions.R
\name{dsd_plot}
\alias{dsd_plot}
\title{Plot disdrometer data}
\usage{
dsd_plot(x, type = "Thies", model = c("Atlas", "Uplinger", "VanDijk"),
  contour = FALSE, theme = "color")
}
\arguments{
\item{x}{A particle size velocity distribution (PSVD) matrix.}

\item{type}{Character vector designing the type of disdrometer,
currently one of 'Thies' or 'Parsivel'. Defaults to
'Thies'.}

\item{model}{Vector. Which theoretical models of V vs. DS curves to
plot on top of the PSVD. Defaults to c('Atlas',
'Uplinger','VanDijk')}

\item{contour}{Logical: should 2d density estimate contour lines be
added to the plot? Defaults to FALSE.}

\item{theme}{Character vector indicating a plotting theme to use.
Current options are 'color' (default) or 'bw' (black and
white).}
}
\value{
A DSD plot.
}
\description{
Produce a PSVD plot: particle count velocity vs. size.
}
\section{References}{
NA
}

\examples{

f <- system.file('extdata', 'rawDataThies', package='disdRo')
files <- list.files(f, '.txt', full.names=TRUE, recursive=TRUE)
dsd <- dsd_read(files, type='Thies')
day <- apply(dsd, c(2,3), sum)
head(day)
# full plot
dsd_plot(day)
# choose one model
dsd_plot(day, model='VanDijk')
# no model
dsd_plot(day, model=NA)
# no model, add contour lines
dsd_plot(day, model=NA, contour=TRUE)

}
