% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disdRo_functions.R
\name{dsd_read}
\alias{dsd_read}
\title{Read raw disdrometer data}
\usage{
dsd_read(files, type = "Thies")
}
\arguments{
\item{files}{A list of files to be processed.}

\item{type}{Character vector designing the type of disdrometer,
currently one of 'Thies' or 'Parsivel' Defaults to
'Thies'.}
}
\value{
An array containing, for each minute, a particle size velocity
distribution (PSVD) matrix, i.e. a matrix of velocity (rows) vs.
size (columns) particle counts.
}
\description{
Retrieves (usually minutal) particle size and velocity distributio PSVD data,
that is the matrix of particle counts arranged by size and velocity classes,
from a list of raw disdrometer data files.
}
\section{References}{
NA
}

\examples{

f <- system.file('extdata', 'rawDataThies', package='disdRo')
files <- list.files(f, '.txt', full.names=TRUE, recursive=TRUE)
dsd <- dsd_read(files, type='Thies')
dim(dsd)
dimnames(dsd)

}
