% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disdRo_functions.R
\name{psvd_model}
\alias{psvd_model}
\title{Terminal drop velocity theoretical models}
\usage{
psvd_model(size, model = "Beard", eta = 1.818e-05, P = 101.325,
  T = 288.15, rho = 1.225, alt = 0)
}
\arguments{
\item{size}{Numeric. A vector of drop sizes (in mm).}

\item{model}{Character. Name of the terminal velocity model to use.
One 'Beard', 'Atlas', 'Uplinger' or 'VanDijk'. Defaults
to 'Beard'.}

\item{eta}{Numeric. A vector of air dynamic viscosity values.
Deafults to 1.818e-04 kg m-1 s-1. Optional, used only
in the Beard model.}

\item{P}{Numeric. A vector of air pressure values. Defaults to
101.325 kPa. Optional, used  only in the Beard model.}

\item{T}{Numeric. A vector of air temperature values. Defaults
to 288.15 K. Optional, used  only in the Beard model.}

\item{rho}{Numeric. A vector of air densities. Defaults to
1.225 kg m-3. Optional, used  only in the Beard and
Atlas models.}

\item{alt}{Numeric. Elevation, which can be used if provided to
estimate pressure, density and temperature of air in
according to the International Standard Atmosphere.
Defaults to 0 m above mean sea level. Optional, used 
in the Beard and Atlas models to estimate parameters
`eta`, `P`, `T` and `rho` if no values are provided.}
}
\value{
A vector of terminal fall velocities. It will cointain NA for any
drop sizes lower than 0.019 mm or larger than 7 mm (Beard model), or lower
than 0.1 and larger than 7 mm for the approximations.
}
\description{
This function implements the Beard model of terminal drop velocity. This
is a physical approximation accounting for different factors such as viscous
drag, drop oblateness, etcetera. The function implemented here is valid for
drop sizes between 0.019 mm (19 µm) up to 7 mm. The function also contains
the approximations of Atlas, Uplinger and Van Dijk. The Beard model is
implemented by default assuming a standard atmosphere: P=101.325 kPa,
T=288.15 K, and rho= 1.225 kg/m3, but other values can be provided.
}
\examples{

size <- seq(0, 6, 0.2)
psvd_model(size)
psvd_model(size, alt=2000)
psvd_model(size, model='Atlas')
psvd_model(size, model='Uplinger')
psvd_model(size, model='VanDijk')

}
