% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disdRo_functions.R
\name{psvd_plot}
\alias{psvd_plot}
\title{Plot disdrometer data}
\usage{
psvd_plot(x, type = "Thies", model = NULL, contour = FALSE,
  theme = "color", filter = NULL, alpha = 0.25)
}
\arguments{
\item{x}{A particle size velocity distribution (PSVD) matrix.}

\item{type}{Character vector designing the type of disdrometer,
currently one of 'Thies' or 'Parsivel'. Defaults to
'Thies'.}

\item{model}{Vector. Which theoretical models of V vs. DS curves to
plot on top of the PSVD. One or more of c('Beard',
'Atlas', Uplinger','VanDijk'). Defaults to 'Beard'.}

\item{contour}{Logical: should 2d density estimate contour lines be
added to the plot? Defaults to FALSE.}

\item{theme}{Character vector indicating a plotting theme to use.
Current options are 'color' (default) or 'bw' (black and
white).}

\item{filter}{A value between 0 and 1. Removes outlier bins, i.e.
those that are between (1-value) and (1+value) far from
the Bear theoretical fall velocity model. Defaults to
Inf (no outliers are removed).}

\item{alpha}{Numeric. Transparency of the filtered (removed) bins.
Defaults to 0.25 (a value of 0 will remove them
completely).}
}
\value{
A PSVD plot.
}
\description{
Produce a PSVD plot: particle count velocity vs. size.
}
\section{References}{
NA
}

\examples{

f <- system.file('extdata', 'rawDataThies', package='disdRo')
files <- list.files(f, '.txt', full.names=TRUE, recursive=TRUE)
dsd <- psvd_read(files, type='Thies')
day <- apply(dsd, c(2,3), sum)
head(day)
# full plot
psvd_plot(day)
# choose one model
psvd_plot(day, model='Beard')
# no model, add contour lines
psvd_plot(day, model=NA, contour=TRUE)

# Use a filter to remove outlier bins
psvd_plot(day, filter=psvd_filter(type='Thies', d=c(0.3,7), tau=0.5))
psvd_plot(day, filter=psvd_filter(type='Thies', d=c(0.3,7), tau=0.5), alpha=0)

}
