% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disdRo_functions.R
\name{psvd_read}
\alias{psvd_read}
\title{Read raw disdrometer data}
\usage{
psvd_read(files, type = "Thies")
}
\arguments{
\item{files}{A list of files to be processed.}

\item{type}{Character vector designing the type of disdrometer,
currently one of 'Thies' or 'Parsivel' Defaults to
'Thies'.}
}
\value{
An array containing, for each minute, a particle size velocity
distribution (PSVD) matrix, i.e. a matrix of size (rows) vs.
velocity (columns) particle counts.
}
\description{
Retrieves (usually minutal) particle size and velocity distribution (PSVD)
data, that is the matrix of particle counts arranged by size and velocity
bins, from a list of raw disdrometer data files.
}
\examples{

f <- system.file('extdata', 'rawDataThies', package='disdRo')
files <- list.files(f, '.txt', full.names=TRUE, recursive=TRUE)
dsd <- psvd_read(files, type='Thies')
dim(dsd)
dimnames(dsd)

}
