% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disdRo_functions.R
\name{pvd_plot}
\alias{pvd_plot}
\title{Particle velocity distribution plot}
\usage{
pvd_plot(x, type = "Thies", filter = NULL, a = NULL)
}
\arguments{
\item{x}{A particle size velocity distribution (PSVD) matrix.}

\item{type}{Character vector designing the type of disdrometer,
currently one of 'Thies' or 'Parsivel'. Defaults to
'Thies'.}

\item{filter}{A value between 0 and 1. Removes outlier bins, i.e.
those that are between (1-value) and (1+value) far from
the Beard theoretical fall velocity model. Defaults to
Inf (no outliers are removed).}

\item{a}{Optional, the sensor area (in cm2). Defaults to the
mean sensor area of Thies and Parsivel disdrometers.}
}
\value{
A PSD plot.
}
\description{
Produce a PSV plot: particle density (number of drops per m^3 of air and
mm of rain) vs particle velocity.
}
\section{References}{
NA
}

\examples{

f <- system.file('extdata', 'rawDataThies', package='disdRo')
files <- list.files(f, '.txt', full.names=TRUE, recursive=TRUE)
dsd <- psvd_read(files, type='Thies')
day <- apply(dsd, c(2,3), sum)
pvd_plot(day)
pvd_plot(day, filter=psvd_filter(type='Thies', d=c(0.3,7), tau=0.5))

}
